Enabled and fixed flake8/pylint checking for zhmccli, and other stuff.
Details for flake8 and pylint checking:
- Added the zhmccli files to the list of files to be checked,
  in the Makefile.
- Fixed any flake8 issues.
- Fixed some pylint issues, one of which revealed the fixes mentioned
  below.

Details for fixes:
- Fixed an error with incorrectly named local variable in
  `_find_partition()`.
- Fixed an error with incorrectly named property
  `Session._password_callback()`.

Details for reorganizations:
- Simplified the way the row and headers are built in
  `print_resources_as_table()`.
- Moved the `cmd_*()` functions from zhmccli.py into the respective
  _cmd_*.py modules.
- Moved `abort_if_false()` from zhmccli.py to _helper.py.
- Renamed `print_properties_in_table()` to `print_properties_as_table()`,
  for consistency.
- Renamed `print_list_in_table()` to `print_resources_as_table()`,
  for consistency.
