+++ Makefile
@@ -29,6 +29,9 @@ endif
 # Name of this Python package (top-level Python namespace + Pypi package name)
 package_name := zhmcclient
 
+# Name of the Python namespace for the CLI
+cli_package_name := zhmccli
+
 # Package version (full version, including any pre-release suffixes, e.g. "0.1.0-alpha1")
 package_version := $(shell python -c "import sys, $(package_name); sys.stdout.write($(package_name).__version__)")
 
@@ -82,6 +85,7 @@ pylint_rc_file := .pylintrc
 check_py_files := \
     setup.py \
     $(wildcard $(package_name)/*.py) \
+    $(wildcard $(cli_package_name)/*.py) \
     $(wildcard tests/unit/*.py) \
     $(wildcard tests/function/*.py) \
     $(wildcard docs/notebooks/*.py) \
@@ -101,6 +105,7 @@ dist_dependent_files := \
     requirements.txt \
     $(wildcard *.py) \
     $(wildcard $(package_name)/*.py) \
+    $(wildcard $(cli_package_name)/*.py) \
 
 # No built-in rules needed:
 .SUFFIXES:
