+++ zhmccli/zhmccli.py
@@ -14,21 +14,14 @@
 from __future__ import absolute_import
 
 import os
-import sys
-import click
-import click_spinner
-from click_repl import register_repl, repl
 import requests.packages.urllib3
+import click
+from click_repl import register_repl, repl
 
-from ._cmd_info import *
-from ._cmd_session import *
-from ._cmd_cpc import *
-from ._cmd_lpar import *
-from ._cmd_partition import *
-from ._helper import *
+from ._helper import CmdContext
 
 
-# TODO: Move functions for sub-commands from zhmccli.py to _cmd_*.py
+# TODO: Documentation for zhmc cli
 # TODO: Automated tests for zhmc cli
 # TODO: Find a way to clarify in the sub-command help texts that the global
 #       options can be used.
@@ -41,11 +34,6 @@ DEFAULT_OUTPUT_FORMAT = 'table'
 DEFAULT_TIMESTATS = False
 
 
-def abort_if_false(ctx, param, value):
-    if not value:
-        raise click.ClickException("Aborted.")
-
-
 @click.group(invoke_without_command=True)
 @click.option('-h', '--host', type=str, envvar='ZHMC_HOST',
               help="Hostname or IP address of the HMC "
@@ -55,7 +43,7 @@ def abort_if_false(ctx, param, value):
                    "(Default: ZHMC_USERID environment variable).")
 @click.option('-o', '--output-format', type=click.Choice(['table', 'json']),
               help='Output format (Default: {of}).'
-                   .format(of=DEFAULT_OUTPUT_FORMAT))
+              .format(of=DEFAULT_OUTPUT_FORMAT))
 @click.option('-t', '--timestats', type=str, is_flag=True,
               help='Show time statistics of HMC operations.')
 @click.version_option(help="Show the version of this command and exit.")
@@ -196,179 +184,5 @@ def cli(ctx, host, userid, output_format, timestats):
     if ctx.invoked_subcommand is None:
         repl(ctx)
 
-
-@cli.command('info')
-@click.pass_obj
-def info(cmd_ctx):
-    """
-    Show information about the HMC.
-    """
-    cmd_ctx.execute_cmd(lambda: cmd_info(cmd_ctx))
-
-
-@cli.group('session')
-def session():
-    """Command group for managing sessions."""
-
-
-@session.command('create')
-@click.pass_obj
-def session_create(cmd_ctx):
-    """Create a HMC session."""
-    cmd_ctx.execute_cmd(lambda: cmd_session_create(cmd_ctx))
-
-
-@session.command('delete')
-@click.pass_obj
-def session_delete(cmd_ctx):
-    """Delete a HMC session."""
-    cmd_ctx.execute_cmd(lambda: cmd_session_delete(cmd_ctx))
-
-
-@cli.group('cpc')
-def cpc():
-    """Command group for managing CPCs."""
-
-@cpc.command('list')
-@click.pass_obj
-def cpc_list(cmd_ctx):
-    """List the CPCs."""
-    cmd_ctx.execute_cmd(lambda: cmd_cpc_list(cmd_ctx))
-
-@cpc.command('show')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.pass_obj
-def cpc_show(cmd_ctx, cpc_name):
-    """Show details of a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_cpc_show(cmd_ctx, cpc_name))
-
-
-@cli.group('lpar')
-def lpar():
-    """Command group for managing LPARs."""
-
-@lpar.command('list')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.pass_obj
-def lpar_list(cmd_ctx, cpc_name):
-    """List the LPARs in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_lpar_list(cmd_ctx, cpc_name))
-
-@lpar.command('show')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('LPAR-NAME', type=str, metavar='LPAR-NAME')
-@click.pass_obj
-def lpar_show(cmd_ctx, cpc_name, lpar_name):
-    """Show details of an LPAR in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_lpar_show(cmd_ctx, cpc_name, lpar_name))
-
-@lpar.command('activate')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('LPAR-NAME', type=str, metavar='LPAR-NAME')
-@click.pass_obj
-def lpar_activate(cmd_ctx, cpc_name, lpar_name):
-    """Activate an LPAR in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_lpar_activate(cmd_ctx, cpc_name,
-                                                  lpar_name))
-
-@lpar.command('deactivate')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('LPAR-NAME', type=str, metavar='LPAR-NAME')
-@click.option('--yes', is_flag=True, callback=abort_if_false,
-              expose_value=False,
-              help='Skip prompt to confirm deactivation of the LPAR.',
-              prompt='Are you sure you want to deactivate the LPAR ?')
-@click.pass_obj
-def lpar_deactivate(cmd_ctx, cpc_name, lpar_name):
-    """Deactivate an LPAR in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_lpar_deactivate(cmd_ctx, cpc_name,
-                                                    lpar_name))
-
-@lpar.command('load')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('LPAR-NAME', type=str, metavar='LPAR-NAME')
-@click.argument('LOAD-ADDRESS', type=str, metavar='LOAD-ADDRESS')
-@click.pass_obj
-def lpar_load(cmd_ctx, cpc_name, lpar_name, load_address):
-    """Load an LPAR in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_lpar_load(cmd_ctx, cpc_name, lpar_name,
-                                              load_address))
-
-
-@cli.group('partition')
-def partition():
-    """Command group for managing partitions."""
-
-@partition.command('list')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.pass_obj
-def partition_list(cmd_ctx, cpc_name):
-    """List the partitions in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_partition_list(cmd_ctx, cpc_name))
-
-@partition.command('create')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.option('--name', type=str, required=True,
-              help='The name of the partition.')
-@click.option('--description', type=str, required=False,
-              help='The description associated with this partition.')
-@click.option('--cp-processors', type=int, required=True,
-              help='Defines the number of general purpose processors (CP).')
-@click.option('--initial-memory', type=int, required=True,
-              help='The initial amount of memory when the partition is '
-                   'started.')
-@click.option('--maximum-memory', type=int, required=True,
-              help='The maximum size while the partition is running.')
-@click.option('--processor-mode', type=str, required=True,
-              help='Defines how processors are allocated to the partition.')
-@click.option('--boot-device', type=str, required=True,
-              help='The type of device from which the partition is booted.')
-@click.pass_context
-def partition_create(ctx, cpc_name, **options):
-    """Create a partition in a CPC."""
-    properties = options_to_properties(options)
-    cmd_ctx = ctx.obj
-    cmd_ctx.execute_cmd(lambda: cmd_partition_create(cmd_ctx, cpc_name,
-                                                     properties))
-
-@partition.command('show')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('PARTITION-NAME', type=str, metavar='PARTITION-NAME')
-@click.pass_obj
-def partition_show(cmd_ctx, cpc_name, partition_name):
-    """Show the details of a partition in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_partition_show(cmd_ctx, cpc_name,
-                                                   partition_name))
-
-@partition.command('start')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('PARTITION-NAME', type=str, metavar='PARTITION-NAME')
-@click.pass_obj
-def partition_start(cmd_ctx, cpc_name, partition_name):
-    """Start a partition in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_partition_start(cmd_ctx, cpc_name,
-                                                    partition_name))
-
-@partition.command('stop')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('PARTITION-NAME', type=str, metavar='PARTITION-NAME')
-@click.pass_obj
-def partition_stop(cmd_ctx, cpc_name, partition_name):
-    """Stop a partition in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_partition_stop(cmd_ctx, cpc_name,
-                                                   partition_name))
-
-@partition.command('delete')
-@click.argument('CPC-NAME', type=str, metavar='CPC-NAME')
-@click.argument('PARTITION-NAME', type=str, metavar='PARTITION-NAME')
-@click.option('--yes', is_flag=True, callback=abort_if_false,
-              expose_value=False,
-              help='Skip prompt to confirm deletion of the partition.',
-              prompt='Are you sure you want to delete this partition ?')
-@click.pass_obj
-def partition_delete(cmd_ctx, cpc_name, partition_name):
-    """Delete a partition in a CPC."""
-    cmd_ctx.execute_cmd(lambda: cmd_partition_delete(cmd_ctx, cpc_name,
-                                                     partition_name))
 
 register_repl(cli)
